requirements.txt
altair==5.0.1
attrs==24.2.0
backports.zoneinfo==0.2.1
blinker==1.6.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
decorator==5.1.1
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
interchange==2021.0.4
Jinja2==3.1.6
jsonschema==4.17.3
markdown-it-py==2.2.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
numpy==1.21.6
packaging==23.2
pandas==1.3.5
pansi==2024.11.0
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
protobuf==4.24.4
py2neo==2021.2.4
pyarrow==12.0.1
pydeck==0.8.1b1
Pygments==2.17.2
Pympler==1.1
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2025.1
pytz-deprecation-shim==0.1.0.post0
pywin32==308
requests==2.31.0
rich==13.8.1
six==1.17.0
smmap==5.0.2
streamlit==1.23.1
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.2
typing_extensions==4.7.1
tzdata==2025.2
tzlocal==4.3.1
urllib3==2.0.7
validators==0.20.0
watchdog==3.0.0
zipp==3.15.0
